requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
langchain==0.1.0
browser-use==0.3.0
google-generativeai==0.3.2
langchain-google-genai==0.1.0
